added support for mysql master/slave setup

+++ lib/appliances/mysql.rb
@@ -2,18 +2,24 @@ Capistrano::Configuration.instance(:must_exist).load do
   task :mysql do
     needs :database
     needs :database_backup
+    needs :mysql_slave_config if current_node.options[:mysql_slave]
   end
   
   task :database do
-    assure(:file, "/etc/mysql/5.0/my.cnf", File.read(mysql_conf_file)) if exists?(:mysql_conf_file)
+    assure(:file, "/etc/mysql/5.0/my.cnf", render(mysql_conf_file, {:mysql_server_id => current_node.options[:mysql_server_id]})) if exists?(:mysql_conf_file)
     assure(:package, "SUNWmysql5")
     svc.setprop("svc:/application/database/mysql:version_50", "mysql/enable_64bit", "true")
     svc.enable("svc:/application/database/mysql:version_50")
     sleep(5) # give mysql server time to start up (service is online before mysql is fully started - especially InnoDB takes a couple seconds)
     pfexec("/usr/mysql/bin/mysqladmin -u root password #{mysql_root_password} || exit 0") if exists?(:mysql_root_password)
+    sql = "grant replication slave, replication client on *.* to '#{db_user}'@'%' identified by '#{db_password}' ;"
+    invoke_command("/usr/mysql/bin/mysql -u root --password=#{mysql_root_password} --execute \"#{sql}\"")
   end
   
   task :database_backup do
+    ftp_backup_host = current_node.options[:ftp_backup_host]
+    ftp_user = current_node.options[:ftp_user]
+    ftp_password = current_node.options[:ftp_password]
     if ftp_backup_host && ftp_user && ftp_password
       logger.info("Setting up daily mysqldump backup for database '#{db_name}' to FTP server '#{ftp_backup_host}' with user '#{ftp_user}'")
       assure :file, "/export/home/#{application_user}/.netrc", "machine #{ftp_backup_host} login #{ftp_user} password #{ftp_password}", :mode => 600
@@ -30,10 +36,17 @@ Capistrano::Configuration.instance(:must_exist).load do
       EOSH
       assure :file, "/export/home/#{application_user}/backup.sh", backup_sh, :mode => 755
     
-      assure :file, "my_crontab", "13 3 * * * /export/home/#{application_user}/backup.sh > /export/home/#{application_user}/mysql_backup.log 2>&1"
-      invoke_command("crontab my_crontab; rm my_crontab")
+      if current_node.options[:primary]
+        assure :file, "my_crontab", "13 3 * * * /export/home/#{application_user}/backup.sh >> /export/home/#{application_user}/mysql_backup.log 2>&1"
+        invoke_command("crontab my_crontab; rm my_crontab")
+      end
     else
-      logger.warn("Database backup not setup. Please set 'ftp_backup_host', 'ftp_user', and 'ftp_password' in your Capfile!")
+      puts("Database backup not setup. Please set 'ftp_backup_host', 'ftp_user', and 'ftp_password' in your Capfile!")
     end
+  end
+  
+  task :mysql_slave_config do
+    sql = "CHANGE MASTER TO MASTER_HOST='#{current_node.options[:mysql_master_host]}', MASTER_USER='#{db_user}', MASTER_PASSWORD='#{db_password}' ;"
+    invoke_command("/usr/mysql/bin/mysql -u root --password=#{mysql_root_password} --execute \"#{sql}\"")
   end
 end
